refactor: Game command 입력에 대한 검증을 Validator에서 BridgeGame으로 옮김
- game command 입력에 대한 책임을 Bridge Game이 지도록 함
