+++ src/main/java/bridge/Model/BridgeGame.java
@@ -42,6 +42,22 @@ public class BridgeGame {
         movementCount += 1;
     }
 
+    public boolean validateGameCommand(String command) {
+        try {
+            validateIsRightCommand(command);
+        } catch (IllegalArgumentException e) {
+            System.out.println(e.getMessage());
+            return false;
+        }
+        return true;
+    }
+
+    void validateIsRightCommand(String command) throws IllegalArgumentException {
+        if (!command.equals("R") && !command.equals("Q")) {
+            throw new IllegalArgumentException(ExceptionMessage.GAME_COMMAND_EXCEPTION_MESSAGE.getExceptionMessage());
+        }
+    }
+
     void retry() {
         round += 1;
         movementCount = 0;
