+++ src/main/java/bridge/View/InputView.java
@@ -5,11 +5,6 @@ import camp.nextstep.edu.missionutils.Console;
  * 사용자로부터 입력을 받는 역할을 한다.
  */
 public class InputView {
-    private final Validator validator;
-
-    public InputView(){
-        validator = new Validator();
-    }
 
     public String readBridgeSize() {
         System.out.println(Message.START_GAME_MESSAGE.getMessage());
@@ -24,10 +19,6 @@ public class InputView {
 
     public String readGameCommand() {
         System.out.println(Message.INPUT_GAME_COMMAND_MESSAGE.getMessage());
-        String gameCommand = Console.readLine();
-        if(!validator.validateGameCommand(gameCommand)){
-            return readGameCommand();
-        }
-        return gameCommand;
+        return Console.readLine();
     }
 }
